refactor: [STEP05-피드백 반영] 어플리케이션 계층의 추상화 삭제(과도한 추상화)
